+++ examples/src/todomvc-sync-electric/src/app/routes/$.tsx
@@ -1,5 +1,0 @@
-import React from 'react'
-
-const Route = () => <></>
-
-export default Route