+++ examples/src/todomvc-sync-electric/src/routes/__root.tsx
@@ -1,0 +1,69 @@
+import 'todomvc-app-css/index.css'
+
+import { LiveStoreProvider } from '@livestore/react'
+import { makeAdapter } from '@livestore/web'
+import LiveStoreSharedWorker from '@livestore/web/shared-worker?sharedworker'
+import { createRootRoute, HeadContent, Outlet, Scripts } from '@tanstack/react-router'
+import * as React from 'react'
+import { unstable_batchedUpdates as batchUpdates } from 'react-dom'
+import { ErrorBoundary } from 'react-error-boundary'
+
+import { schema } from '@/livestore/schema.js'
+
+import LiveStoreWorker from '../livestore/livestore.worker?worker'
+
+const RootComponent = () => {
+  const storeId = getStoreId()
+  const adapter = makeAdapter({
+    storage: { type: 'opfs' },
+    worker: LiveStoreWorker,
+    sharedWorker: LiveStoreSharedWorker,
+  })
+
+  return (
+    <RootDocument>
+      <ErrorBoundary fallback={<div>Something went wrong</div>}>
+        <LiveStoreProvider
+          schema={schema}
+          storeId={storeId}
+          renderLoading={() => <div>Loading...</div>}
+          adapter={adapter}
+          batchUpdates={batchUpdates}
+        >
+          <Outlet />
+        </LiveStoreProvider>
+      </ErrorBoundary>
+    </RootDocument>
+  )
+}
+
+const RootDocument = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <html>
+      <head>
+        <HeadContent />
+      </head>
+      <body>
+        {children}
+        <Scripts />
+      </body>
+    </html>
+  )
+}
+
+export const Route = createRootRoute({
+  component: RootComponent,
+})
+
+const getStoreId = () => {
+  if (typeof window === 'undefined') return 'unused'
+
+  const searchParams = new URLSearchParams(window.location.search)
+  const storeId = searchParams.get('storeId')
+  if (storeId !== null) return storeId
+
+  const newAppId = crypto.randomUUID()
+  searchParams.set('storeId', newAppId)
+
+  window.location.search = searchParams.toString()
+}