+++ examples/src/todomvc-sync-electric/src/routes/index.tsx
@@ -1,17 +1,11 @@
-import 'todomvc-app-css/index.css'
-
-import type { MetaFunction } from '@remix-run/node'
+import { createFileRoute } from '@tanstack/react-router'
 import React from 'react'
 
 import { Footer } from '@/components/Footer.js'
 import { Header } from '@/components/Header.js'
 import { MainSection } from '@/components/MainSection.js'
 
-export const meta: MetaFunction = () => {
-  return [{ title: 'New Remix App' }, { name: 'description', content: 'Welcome to Remix!' }]
-}
-
-export const Route = () => {
+const Home = () => {
   return (
     <div>
       <section className="todoapp">
@@ -23,4 +17,6 @@ export const Route = () => {
   )
 }
 
-export default Route
+export const Route = createFileRoute('/')({
+  component: Home,
+})