+++ examples/src/todomvc-sync-electric/src/ssr.tsx
@@ -1,0 +1,6 @@
+import { getRouterManifest } from '@tanstack/start/router-manifest'
+import { createStartHandler, defaultStreamHandler } from '@tanstack/start/server'
+
+import { createRouter } from './router.js'
+
+export default createStartHandler({ createRouter, getRouterManifest })(defaultStreamHandler)